move ReplyHandleWrapper out of testing and into types
