+++ crack-core/src/lib.rs
@@ -76,7 +76,7 @@ pub type FrameworkContext<'a> = poise::FrameworkContext<'a, Data, CommandError>;
 
 use crate::messaging::message::CrackedMessage;
 use crate::serenity::prelude::SerenityError;
-use crack_testing::reply_handle_trait::MessageOrReplyHandle;
+use crack_types::reply_handle::MessageOrReplyHandle;
 
 impl From<CrackedError> for SerenityError {
     fn from(_e: CrackedError) -> Self {
