+++ crack-types/Cargo.toml
@@ -35,6 +35,7 @@ reqwest = { version = "0.12", default-features = false, features = [
     "macos-system-configuration",
 ] }
 url = "2.5"
+thiserror = "1.0"
 
 [dependencies.rspotify]
 version = "0.13"
